feat: Refactor useMeteorological hook
This commit refactors the useMeteorological hook by renaming the dataHM variable to data and adding the getMeteorologicalData function. The dataHM variable is now named data to improve clarity and consistency. Additionally, the getMeteorologicalData function is added to provide a more descriptive name for retrieving meteorological data. These changes enhance the readability and maintainability of the code.